+++ src/Controller/AbstractResourceController.php
@@ -2,13 +2,12 @@
 
 namespace Resourceful\Controller;
 use Silex\Application;
+use \Resourceful\StoreHelpers\StoreHelpers;
 
 
 class AbstractResourceController
 {
     protected $schema;
-
-    use \Resourceful\StoreHelpers\StoreHelpers;
 
     public function __construct($schema=null)
     {
@@ -17,8 +16,6 @@ class AbstractResourceController
 
 	protected function getStore(Application $app)
     {
-    	$type = $this->schema?basename($this->schema):'schema';
-		
-		return $this->getStoreForType($type, $app);
+		return StoreHelpers::getStoreForSchema($this->schema, $app);
     }
 }